+++ src/methods/index.js
@@ -1,0 +1,29 @@
+import _findTarget      from './_findTarget.js'
+import _listen         from './_listen.js'
+import _readKeyEvent   from './_readKeyEvent.js'
+import _readMouseEvent from './_readMouseEvent.js'
+import _readShortcut   from './_readShortcut.js'
+import _specialChars   from './_specialChars.js'
+
+import load           from './load.js'
+import unload         from './unload.js'
+import changeContext  from './changeContext.js'
+import listShortcuts  from './listShortcuts.js'
+
+
+
+export default {
+// Internal methods
+    _findTarget
+  , _listen
+  , _readKeyEvent
+  , _readMouseEvent
+  , _readShortcut
+  , _specialChars
+
+// Public methods
+  , changeContext
+  , listShortcuts
+  , load
+  , unload
+}